Alterando tipo de qtd para Integer

+++ src/main/java/com/jacyirice/pw2/aula11/models/entity/ItemVenda.java
@@ -29,7 +29,7 @@ public class ItemVenda  implements Serializable{
     private Integer id;
 //    @Min(value=1, message = "oi")
     @NotNull
-    private Double qtd;
+    private Integer qtd;
 
     @ManyToOne
     @JoinColumn(name = "id_produto")
@@ -46,16 +46,16 @@ public class ItemVenda  implements Serializable{
     public void setId(Integer id) {
         this.id = id;
     }
-    
-    public Double getQtd() {
+
+    public Integer getQtd() {
         return qtd;
     }
 
-    public void setQtd(Double qtd) {
+    public void setQtd(Integer qtd) {
         this.qtd = qtd;
     }
     
-    public void addQtd(Double qtd) {
+    public void addQtd(Integer qtd) {
         this.qtd += qtd;
     }
